Update io.h
Fix typo caused by copy-paste :^)

+++ kernel/src/io.h
@@ -7,7 +7,7 @@ void out8  (u16 port, u8 value);
 u8   in8   (u16 port);
 void out16 (u16 port, u16 value);
 u16  in16  (u16 port);
-void out32 (u16 port, u16 value);
+void out32 (u16 port, u32 value);
 u32  in32  (u16 port);
 
 void io_wait();
